Move the joint limit struct to the joint limiter struct and added base interface class
