+++ hardware_interface/src/resource_manager.cpp
@@ -34,7 +34,7 @@
 #include "hardware_interface/system.hpp"
 #include "hardware_interface/system_interface.hpp"
 #include "hardware_interface/types/hardware_interface_type_values.hpp"
-#include "joint_limits/joint_limits_struct.hpp"
+#include "joint_limits/joint_limiter_struct.hpp"
 #include "lifecycle_msgs/msg/state.hpp"
 #include "pluginlib/class_loader.hpp"
 #include "rcutils/logging_macros.h"
